+++ src/main/java/kitchenpos/order/application/OrderValidator.java
@@ -1,0 +1,7 @@
+package kitchenpos.order.application;
+
+import kitchenpos.table.domain.TableGroup;
+
+public interface OrderValidator {
+	void validateUngroup(TableGroup tableGroup);
+}
