+++ src/main/java/kitchenpos/table/application/TableGroupService.java
@@ -2,15 +2,12 @@ package kitchenpos.table.application;
 
 import static kitchenpos.table.ui.request.TableGroupRequest.*;
 
-import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import kitchenpos.order.domain.OrderRepository;
-import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.order.domain.OrderTable;
 import kitchenpos.table.domain.OrderTableRepository;
 import kitchenpos.table.domain.TableGroup;
@@ -21,17 +18,17 @@ import kitchenpos.table.ui.response.TableGroupResponse;
 @Service
 public class TableGroupService {
 	private final TableGroupRepository tableGroupRepository;
-	private final OrderRepository orderRepository;
 	private final OrderTableRepository orderTableRepository;
+	private final TableValidator tableValidator;
 
 	public TableGroupService(
 		final TableGroupRepository tableGroupRepository,
-		final OrderRepository orderRepository,
-		final OrderTableRepository orderTableRepository
+		final OrderTableRepository orderTableRepository,
+		final TableValidator tableValidator
 	) {
 		this.tableGroupRepository = tableGroupRepository;
-		this.orderRepository = orderRepository;
 		this.orderTableRepository = orderTableRepository;
+		this.tableValidator = tableValidator;
 	}
 
 	@Transactional
@@ -43,7 +40,7 @@ public class TableGroupService {
 	@Transactional
 	public void ungroup(final Long tableGroupId) {
 		final TableGroup tableGroup = tableGroup(tableGroupId);
-		validate(tableGroup);
+		tableValidator.validateUngroup(tableGroup);
 		tableGroup.ungroup();
 	}
 
@@ -60,18 +57,6 @@ public class TableGroupService {
 
 	private OrderTable orderTable(long id) {
 		return orderTableRepository.orderTable(id);
-	}
-
-	private void validate(TableGroup tableGroup) {
-		if (isCookingOrMeal(tableGroup)) {
-			throw new IllegalArgumentException("조리중이거나 식사중인 테이블은 테이블 그룹 해제할 수 없습니다.");
-		}
-	}
-
-	private boolean isCookingOrMeal(TableGroup tableGroup) {
-		return orderRepository.existsByOrderTableIdInAndOrderStatusIn(
-			tableGroup.orderTableIds(),
-			Arrays.asList(OrderStatus.COOKING, OrderStatus.MEAL));
 	}
 
 	private TableGroup tableGroup(Long tableGroupId) {
